Remove pId (only one http request at a time)
Previously, if we clicked on a link and it was still loading, hovering over another link would preload this latter link in parallel.

This added huge (HUGE) complexity to the InstantClick’s equation, now in the rare case that a link is stuck, if we click on another one (or the same) it just redirects the user with location.href. Big win in simplicity, and a win in usability in the case that the page is stuck: clicking on another link gives back the browser’s loading indicator.