Prevent preloading 100 ms after page display
Prevents fetching the link your mouse’s cursor is over after a new page is displayed. The most common case for this is when you’ve clicked on an item that have the same position on the previous and current page.

Thanks to quchen on Reddit for the idea.

This commit also changes p.timing and p.timingStart:
- p.timingStart is now p.timing.start
- p.timing is now p.timing.ready
- p.timing.display is when a page is displayed

Also in this commit: detect in `readystatechange()` if a request is aborted.