Don’t act on mouseover/mousedown if there has been preloading in the last 500ms
Fixes the bug in the last commit, click wouldn’t fire because mouseover/mousedown prevented that, for some reason (haven’t dig).

It introduces a regression though, if you tap near a link, it will not be loaded through InstantClick. Previously, it would load through InstantClick (but be preloaded just 1ms before, on mousedown).

A way to fix that in the future is to preload on click if it hasn’t been done. Currently, no preload on click means InstantClick will redirect to the new page, without pjax. Fixing this may also be useful for those that navigate without their mouses.